+++ treesource/TreeMaker.h
@@ -57,10 +57,10 @@ class TreeMaker: public SubsysReco
   double _b_et_iso_calotower_R04[500];
 
   int _b_particle_n;
-  double _b_particle_pt[1000];
-  double _b_particle_et[1000];
-  double _b_particle_eta[1000];
-  double _b_particle_phi[1000];
+  float _b_particle_pt[1000];
+  float _b_particle_et[1000];
+  float _b_particle_eta[1000];
+  float _b_particle_phi[1000];
   int _b_particle_pid[1000];
